Fix (matchers): Implement pagination for Get Many with standard handleGetAllOperation. Fixes issue where only first 25 matcher results could be retrieved.
